+++ CHANGELOG.md
@@ -59,6 +59,7 @@ This name should be decided amongst the team before the release.
 - [#1139](https://github.com/topiary/topiary/pull/1139) Added `rootcause::Report` handling in topiary-cli.
 - [#1283](https://github.com/topiary/topiary/pull/1283) Render Nickel parsing error diagnostics
 - [#1298] https://github.com/topiary/topiary/pull/1298) Update WIT formatter for tree-sitter-wit to v1.4
+- [#1303] https://github.com/topiary/topiary/pull/1303) Add `bin/verify-documented-usage.sh --no-nix-shell` flag
 
 ### Fixed
 - [#1176](https://github.com/topiary/topiary/pull/1176) Increase the stack size to 4MiB in Windows builds.